partykit: extract room circuit breaker
